rename some fische header to have ".hpp" as end
This done to have better usable in coming changes.
As the fische code by add-on was on begin defined as a "C" code only and over the year some "C++" parts implemented to them.
About this and to become a usable "C" ABI support back to them are the headers renamed and to identify in furture which relates to "C" and whicht to "C++".
