+++ src/fische/analyst.hpp
