+++ src/fische/wavepainter.hpp
